refactor(kernel): break the commands-services dependency cycle
Move the PicoCommand interface from commands/ to a new kernel/ package and
introduce services/ProfileSetupCommandProvider so CommandExecutor names an
abstraction instead of constructing SetupProfile directly. No file under
services/ now references a type in commands/.
