+++ src/main/kotlin/com/rustyrazorblade/easydblab/commands/PicoBaseCommand.kt
@@ -4,6 +4,7 @@ import com.rustyrazorblade.easydblab.Context
 import com.rustyrazorblade.easydblab.configuration.ClusterStateManager
 import com.rustyrazorblade.easydblab.events.EventBus
 import com.rustyrazorblade.easydblab.events.EventContext
+import com.rustyrazorblade.easydblab.kernel.PicoCommand
 import com.rustyrazorblade.easydblab.providers.ssh.RemoteOperationsService
 import org.koin.core.component.KoinComponent
 import org.koin.core.component.inject
